+++ src/Helper.hh
@@ -64,7 +64,7 @@ inline auto get_rank(const Card &card1, const Card &card2,
 
 inline auto get_win_pct(const PreflopCombo &hero, const PreflopCombo &villain,
                         const std::vector<Card> board,
-                        const double accuracy_margin = 0.01) -> double {
+                        const float accuracy_margin = 0.01) -> float {
   assert((board.size() == 3 || board.size() == 4) &&
          "get_win_pct: expected board of size 3 or 4 (undealt turn/river)");
   EquityCalculator eq;
