+++ src/game/Game.hh
@@ -5,8 +5,8 @@ namespace GameParams {
 constexpr int NUM_CARDS{52};
 constexpr int NUM_SUITS{4};
 constexpr std::array<char, 4> suitReverseArray = {'c', 'd', 'h', 's'};
-constexpr std::array<double, 5> BET_SIZES{0.5, 1.0};
-constexpr std::array<double, 1> RAISE_SIZES{0.5};
+constexpr std::array<float, 5> BET_SIZES{0.5, 1.0};
+constexpr std::array<float, 1> RAISE_SIZES{0.5};
 } // namespace GameParams
 
 enum class Street { FLOP = 3, TURN = 4, RIVER = 5 };
